+++ src/lib/index.ts
@@ -3,9 +3,12 @@ import Footer from "./Footer.tsx";
 import DialogOperation from "./DialogOperation.tsx";
 import AccountingDiary from "./AccountingDiary.tsx";
 import AccountingDiaryWrapper from "./AccountingDiaryWrapper.tsx";
+import LedgerView from "./LedgerView.tsx";
 import { GlobalProvider } from "./context.tsx";
+import { useAccountingDiary } from "./useAccountingDiary.ts";
 
-export { Header, Footer, DialogOperation, AccountingDiary, AccountingDiaryWrapper, GlobalProvider };
-export type { IDataItem, IStyleConfig, ILabels, SortField, SortOrder } from '../types/common';
+export { Header, Footer, DialogOperation, AccountingDiary, AccountingDiaryWrapper, LedgerView, GlobalProvider, useAccountingDiary };
+export type { IDataItem, IStyleConfig, ILabels, SortField, SortOrder, ViewMode, AccountingDiaryHandle, UseAccountingDiaryOptions, UseAccountingDiaryReturn } from '../types/common';
 export { defaultLabels } from '../types/common';
+export { exportToCSV, exportToExcel, exportToJSON, importFromCSV, importFromJSON } from './helpers/exportUtils';
 export default AccountingDiaryWrapper;